+++ packages/generator/src/loaders/loadPoShortcodes.ts
@@ -5,7 +5,13 @@ import PO from 'pofile';
 import { LOCALE_COUNTRIES } from '../constants';
 import POManager from '../parsers/POManager';
 
+const instanceCache: Partial<Record<Locale, POManager>> = {};
+
 export default async function loadPoShortcodes(locale: Locale): Promise<POManager> {
+  if (instanceCache[locale]) {
+    return instanceCache[locale]!;
+  }
+
   const poPath = path.resolve(process.cwd(), `po/${locale}/shortcodes.po`);
   const po = PO.parse(await fs.promises.readFile(poPath, 'utf8'));
 
@@ -13,5 +19,10 @@ export default async function loadPoShortcodes(locale: Locale): Promise<POManage
     po.comments = [LOCALE_COUNTRIES[locale]];
   }
 
-  return new POManager(poPath, po);
+  const instance = new POManager(poPath, po);
+
+  // eslint-disable-next-line require-atomic-updates
+  instanceCache[locale] = instance;
+
+  return instance;
 }